FEATURE NeuralVar AIC/BIC + mne logger

+++ nlgc/_nlgc.py
@@ -34,7 +34,7 @@ _default_lambda_range = np.asanyarray([5e-1, 2e-1, 1e-1, 5e-2, 2e-2, 1e-2, 5e-3,
 
 class NLGC:
     def __init__(self, subject, nx, ny, t, p, n_eigenmodes, n_segments, d_raw, bias_f, bias_r,
-            model_f, conv_flag, label_names, label_vertidx, debug=None):
+                 model_f, conv_flag, label_names, label_vertidx, debug=None):
 
         self.subject = subject
         self.nx = nx
@@ -115,38 +115,34 @@ def nlgc_map(name, evoked, forward, noise_cov, labels, order, self_history=None,
         raise ValueError("Length of patch_idx should not be zero")
 
     n, _ = G.shape
-    ex_G = np.zeros((n, len(patch_idx) * n_eigenmodes))
-    for idx, this_patch in enumerate(patch_idx):
-        ex_G[:, idx * n_eigenmodes: (idx + 1) * n_eigenmodes] = \
-            G[:, this_patch * n_eigenmodes: (this_patch + 1) * n_eigenmodes]
-    ROIs = list(range(len(patch_idx)))
-
-    n, nnx = ex_G.shape
-    len_patch_idx = nnx // n_eigenmodes
+    n, nnx = G.shape
+    nx = nnx // n_eigenmodes
     _, t = M.shape
     tt = t // n_segments
 
-    d_raw = np.zeros((n_segments, len_patch_idx, len_patch_idx))
-    bias_r = np.zeros((n_segments, len_patch_idx, len_patch_idx))
+    d_raw = np.zeros((n_segments, nx, nx))
+    bias_r = np.zeros((n_segments, nx, nx))
     bias_f = np.zeros((n_segments, 1))
-    conv_flag = np.zeros((n_segments, len_patch_idx, len_patch_idx))
+    conv_flag = np.zeros((n_segments, nx, nx))
+
     models = []
 
-    for n in range(0, n_segments):
-        print('Segment: ', n + 1)
+    for this_segment in range(0, n_segments):
+        logger.info('Segment: ', this_segment + 1)
         d_raw_, bias_r_, bias_f_, model_f, conv_flag_ = \
-            _gc_extraction(M[:, n * tt: (n + 1) * tt], ex_G, r, p=order, p1=self_history, n_eigenmodes=n_eigenmodes,
-                           ROIs=ROIs,
+            _gc_extraction(M[:, this_segment * tt: (this_segment + 1) * tt], G, r, p=order, p1=self_history,
+                           n_eigenmodes=n_eigenmodes,
+                           ROIs=patch_idx,
                            alpha=alpha, beta=beta, cv=cv, lambda_range=lambda_range, max_iter=max_iter,
                            max_cyclic_iter=max_cyclic_iter, tol=tol, sparsity_factor=sparsity_factor,
                            use_lapack=use_lapack, use_es=use_es, var_thr=var_thr)
-        d_raw[n] = d_raw_
-        bias_r[n] = bias_r_
-        bias_f[n] = bias_f_
+        d_raw[this_segment] = d_raw_
+        bias_r[this_segment] = bias_r_
+        bias_f[this_segment] = bias_f_
         models.append(model_f)
-        conv_flag[n] = conv_flag_
+        conv_flag[this_segment] = conv_flag_
 
-    nlgc_obj = NLGC(name, len_patch_idx, n, t, order, n_eigenmodes, n_segments, d_raw, bias_f, bias_r, models,
+    nlgc_obj = NLGC(name, nx, n, t, order, n_eigenmodes, n_segments, d_raw, bias_f, bias_r, models,
                     conv_flag, label_names, label_vertidx)
 
     return nlgc_obj
@@ -155,7 +151,6 @@ def nlgc_map(name, evoked, forward, noise_cov, labels, order, self_history=None,
 def _gc_extraction(y, f, r, p, p1, n_eigenmodes=2, var_thr=1.0, ROIs=[], alpha=0, beta=0,
         lambda_range=None, max_iter=500, max_cyclic_iter=3,
         tol=1e-5, sparsity_factor=0.0, cv=5, use_lapack=True, use_es=True):
-    logger = logging.getLogger(__name__)
     n, m = f.shape
     nx = m // n_eigenmodes
 
@@ -221,8 +216,6 @@ def _gc_extraction(y, f, r, p, p1, n_eigenmodes=2, var_thr=1.0, ROIs=[], alpha=0
         sorted_pow_ratio /= sorted_pow_ratio[-1]
         idx = ((sorted_pow_ratio > var_thr) != 0).argmax()
         ROIs = sorted_idx[:idx + 1]
-    else:
-        ROIs = list(range(m // n_eigenmodes))
 
     links_to_check = []
     for i, j in itertools.product(ROIs, repeat=2):
@@ -249,22 +242,28 @@ def _gc_extraction(y, f, r, p, p1, n_eigenmodes=2, var_thr=1.0, ROIs=[], alpha=0
     args = (r, lambda_f, a_f, q_f, p, p1, n_eigenmodes, use_lapack)  # can be passed directly
 
     # Parallel
-    n_jobs = min(cpu_count(), len(links_to_check))
-    Parallel(n_jobs=n_jobs, verbose=10)(delayed(_learn_reduced_model_parallel)(link, *(shared_args + args),
-                                                                               **kwargs) for link in links_to_check)
-    # # serial
-    # [_learn_reduced_model_parallel(link, *(shared_args + args), ** kwargs) for link in links_to_check]
+    if len(links_to_check) == 0:
+        bias_f = 0
+    else:
+        n_jobs = min(cpu_count(), len(links_to_check))
+        Parallel(n_jobs=n_jobs, verbose=10)\
+            (delayed(_learn_reduced_model_parallel)(link, *(shared_args + args), **kwargs) for link in links_to_check)
+        # # serial
+        # [_learn_reduced_model_parallel(link, *(shared_args + args), ** kwargs) for link in links_to_check]
 
-    ll_r = np.reshape(shared_ll_r, dev_raw.shape).copy()
-    bias_r = np.reshape(shared_bias_r, dev_raw.shape).copy()
-    conv_flag = np.reshape(shared_conv_flag, dev_raw.shape).copy()
-    for shm in (shm_conv_flag, shm_bias_r, shm_f, shm_ll_r, shm_y):
-        shm.close()
-        shm.unlink()
+        ll_r = np.reshape(shared_ll_r, dev_raw.shape).copy()
+        bias_r = np.reshape(shared_bias_r, dev_raw.shape).copy()
+        conv_flag = np.reshape(shared_conv_flag, dev_raw.shape).copy()
+        for shm in (shm_conv_flag, shm_bias_r, shm_f, shm_ll_r, shm_y):
+            shm.close()
+            try:
+                shm.unlink()
+            except:
+                logger.info(f"Unlink shared-memory issue!")
 
-    indices = tuple(z for z in zip(*links_to_check))
-    dev_raw[indices] = 2 * model_f.ll
-    dev_raw[indices] -= 2 * ll_r[indices]
+        indices = tuple(z for z in zip(*links_to_check))
+        dev_raw[indices] = 2 * model_f.ll
+        dev_raw[indices] -= 2 * ll_r[indices]
 
     # # Old log ratio implementation
     # dev_raw_[j, i] = sum(map(lambda x: np.log(model_r._parameters[2][x, x]) - np.log(model_f._parameters[2][x, x]),
@@ -294,7 +293,6 @@ def _learn_reduced_model(i, j, y, f, r, lambda_f, a, q, n, p, p1, n_eigenmodes, 
 
 def _learn_reduced_model_parallel(link_index, info_y, info_f, info_bias_r, info_ll_r, info_conv_flag, r, lambda_f, a, q,
         p, p1, n_eigenmodes, use_lapack, alpha, beta, **kwargs):
-    logger = logging.getLogger(__name__)
     try:
         y, shm_y = link_share_memory(info_y)
         f, shm_f = link_share_memory(info_f)
